contract-check/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1

langchain==0.3.14
langchain-openai==0.3.1
pydantic==2.10.4


